Did a lot of work with server/client Server -> Client interation now works. Just need do interpolation on the client and fix some bugs on the server

+++ server/src/me/eddiep/ghost/server/network/packet/impl/MatchStatusPacket.java
@@ -17,7 +17,7 @@ public class MatchStatusPacket extends Packet {
 
         Boolean status = (Boolean) args[0];
 
-        write(0x06)
+        write((byte)0x06)
           .write(status)
           .endTCP();
     }
